gen_swupdate.py: fix libconfig include path
Libconfig include directives are broken, because the "includedir" points
by default to the current working directory. gen_swupdate.py switches
the current working directory before parsing the content of the
template. Therefore it tries to open the include files from the
temporary directory, not from the template directory.

Switch to the temporary directory after parsing the template fixes this
issue.

Signed-off-by: Holger Dengler <holger@hdengler.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>